chore(file structure): clean up file structure, move things into separate files
