+++ src/gatsby-browser.js
@@ -1,1 +1,1 @@
-exports.wrapPageElement = require('./lib/wrap-page').default;
+exports.wrapPageElement = require('./lib/browser/wrapPageElement');
