+++ src/lib/browser/wrapPageElement.js
@@ -1,12 +1,12 @@
 const React = require('react');
-const { PREVIEW_CONTEXT } = require('./const');
+const { PREVIEW_CONTEXT } = require('../const');
 const { ApolloProvider } = require('react-apollo-hooks');
 const { IntrospectionFragmentMatcher } = require('apollo-cache-inmemory');
 const ApolloClient = require('apollo-client').default;
 const { InMemoryCache } = require('apollo-cache-inmemory');
 const { HttpLink } = require('apollo-link-http');
-const PreviewUIComponent = require('./components/PreviewUI').default;
-const PreviewFetcher = require('./components/PreviewFetcher').default;
+const PreviewUIComponent = require('../components/PreviewUI').default;
+const PreviewFetcher = require('../components/PreviewFetcher').default;
 
 const isolatedQueries = GATSBY_PLUGIN_GRAPHQL_PREVIEW_ISOLATED_QUERIES;
 const fragmentTypes = GATSBY_PLUGIN_GRAPHQL_PREVIEW_FRAGMENT_TYPES;
@@ -17,7 +17,7 @@ const fragmentMatcher = new IntrospectionFragmentMatcher({
 const cache = new InMemoryCache({ fragmentMatcher });
 
 // eslint-disable-next-line react/prop-types,react/display-name
-exports.default = ({ element, props }, options) => {
+exports = ({ element, props }, options) => {
   const componentId = props.pageContext[PREVIEW_CONTEXT];
   const isolatedQuery = isolatedQueries[componentId];
